Move default Hibernate 2LC implementation to package io.quarkus.hibernate.orm.runtime.cache.local
