+++ extensions/hibernate-orm/runtime/src/main/java/io/quarkus/hibernate/orm/runtime/cache/local/InternalCacheConfig.java
@@ -1,0 +1,10 @@
+package io.quarkus.hibernate.orm.runtime.cache.local;
+
+import java.time.Duration;
+
+final class InternalCacheConfig {
+
+    long objectCount;
+    Duration maxIdle;
+
+}
