+++ extensions/hibernate-orm/runtime/src/main/java/io/quarkus/hibernate/orm/runtime/cache/local/InternalRegion.java
@@ -1,0 +1,23 @@
+package io.quarkus.hibernate.orm.runtime.cache.local;
+
+import java.util.Comparator;
+
+interface InternalRegion {
+
+    boolean checkValid();
+
+    void beginInvalidation();
+
+    void endInvalidation();
+
+    long getLastRegionInvalidation();
+
+    String getName();
+
+    void clear();
+
+    Comparator<?> getComparator(String subclass);
+
+    void addComparator(String name, Comparator<?> comparator);
+
+}
