+++ extensions/hibernate-orm/runtime/src/main/java/io/quarkus/hibernate/orm/runtime/cache/local/Time.java
@@ -1,0 +1,36 @@
+package io.quarkus.hibernate.orm.runtime.cache.local;
+
+import java.time.Duration;
+import java.time.temporal.ChronoUnit;
+
+enum Time {
+    ;
+
+    private static final Duration FOREVER = ChronoUnit.FOREVER.getDuration();
+
+    static Duration forever() {
+        return FOREVER;
+    }
+
+    static boolean isForever(Duration duration) {
+        return FOREVER.equals(duration);
+    }
+
+    @FunctionalInterface
+    interface NanosService {
+        long nanoTime();
+    }
+
+    @FunctionalInterface
+    interface MillisService {
+        MillisService SYSTEM = new MillisService() {
+            @Override
+            public long milliTime() {
+                return System.currentTimeMillis();
+            }
+        };
+
+        long milliTime();
+    }
+
+}
