+++ extensions/hibernate-orm/runtime/src/main/java/io/quarkus/hibernate/orm/runtime/service/QuarkusRegionFactoryInitiator.java
@@ -10,7 +10,8 @@ import org.hibernate.cache.spi.RegionFactory;
 import org.hibernate.cfg.AvailableSettings;
 import org.hibernate.internal.util.config.ConfigurationHelper;
 import org.hibernate.service.spi.ServiceRegistryImplementor;
-import org.infinispan.quarkus.hibernate.cache.QuarkusInfinispanRegionFactory;
+
+import io.quarkus.hibernate.orm.runtime.cache.local.QuarkusInfinispanRegionFactory;
 
 public final class QuarkusRegionFactoryInitiator implements StandardServiceInitiator<RegionFactory> {
