--- extensions/hibernate-orm/runtime/src/main/java/org/infinispan/quarkus/hibernate/cache/InternalCacheConfig.java
@@ -1,10 +1,0 @@
-package org.infinispan.quarkus.hibernate.cache;
-
-import java.time.Duration;
-
-final class InternalCacheConfig {
-
-    long objectCount;
-    Duration maxIdle;
-
-}
