--- extensions/hibernate-orm/runtime/src/main/java/org/infinispan/quarkus/hibernate/cache/InternalRegion.java
@@ -1,23 +1,0 @@
-package org.infinispan.quarkus.hibernate.cache;
-
-import java.util.Comparator;
-
-interface InternalRegion {
-
-    boolean checkValid();
-
-    void beginInvalidation();
-
-    void endInvalidation();
-
-    long getLastRegionInvalidation();
-
-    String getName();
-
-    void clear();
-
-    Comparator<?> getComparator(String subclass);
-
-    void addComparator(String name, Comparator<?> comparator);
-
-}
